scripts/requirements.txt
# Script dependencies
python-dotenv>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
python-docx>=1.1.0
psycopg2-binary>=2.9.0

# Testing dependencies
pytest>=8.3.0
pytest-cov>=5.0.0

